sauvgarde des info faite dans gestion clients puis gestion historique a utiliser la meme chose pour tt les gestionaire

+++ kasbah-terminal/utile.h
@@ -28,7 +28,7 @@ public:
 	};
 
 
-	static long unsigned getIdFromtime() {
+	static unsigned getIdFromtime() {
 		time_t now;
 		time(&now);
 		struct tm* local = localtime(&now);
